1. modify consul support the makefile
