+++ registry/servicediscovery/consul/go-client/cmd/version.go
