+++ registry/servicediscovery/consul/go-server/cmd/version.go
