+++ registry/servicediscovery/consul/go-server/conf/server.yml
