+++ registry/servicediscovery/consul/go-server/docker/docker-compose.yml
@@ -1,0 +1,8 @@
+version: '3'
+
+services:
+  consul:
+    image: consul:latest
+    ports:
+      - 8500:8500
+    restart: on-failure
