--- registry/servicediscovery/consul/go-server/profiles/test/server.yml
@@ -1,79 +1,0 @@
-# dubbo server yaml configure file
-
-
-# application config
-application:
-  organization: "ikurento.com"
-  name: "user-info-server"
-  module: "dubbogo user-info server"
-  version: "0.0.1"
-  owner: "ZX"
-  environment: "dev"
-  metadataType: "local"
-
-
-registries:
-  "demoServiceDiscovery":
-    protocol: "service-discovery"
-    params:
-      service_discovery: "nacos1"
-      name_mapping: "dynamic"
-      metadata: "default"
-
-remote:
-  nacos:
-    address: "127.0.0.1:8848"
-    timeout: "5s"
-
-metadata_report:
-  protocol: "nacos"
-  remote_ref: "nacos"
-
-service_discovery:
-  nacos1:
-    protocol: "nacos"
-    remote_ref: "nacos"
-
-config_center:
-  protocol: "nacos"
-  address: "127.0.0.1:8848"
-
-services:
-  "UserProvider":
-    # 可以指定多个registry，使用逗号隔开;不指定默认向所有注册中心注册
-    registry: "demoServiceDiscovery"
-    protocol: "dubbo"
-    # 相当于dubbo.xml中的interface
-    interface: "com.ikurento.user.UserProvider"
-    loadbalance: "random"
-    warmup: "100"
-    cluster: "failover"
-    methods:
-      - name: "GetUser"
-        retries: 1
-        loadbalance: "random"
-
-protocols:
-  "dubbo":
-    name: "dubbo"
-    port: 20000
-
-
-protocol_conf:
-  dubbo:
-    session_number: 700
-    session_timeout: "180s"
-    getty_session_param:
-      compress_encoding: false
-      tcp_no_delay: true
-      tcp_keep_alive: true
-      keep_alive_period: "120s"
-      tcp_r_buf_size: 262144
-      tcp_w_buf_size: 65536
-      pkg_rq_size: 1024
-      pkg_wq_size: 512
-      tcp_read_timeout: "1s"
-      tcp_write_timeout: "5s"
-      wait_timeout: "1s"
-      max_msg_len: 1024000
-      session_name: "server"
